Split filters into filters.rs from groups.rs
